pyproject.toml
[project]
name = "lightning-hydra"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.9.21"
dependencies = [
    "torch==2.8.0",
    "torchaudio==2.8.0",
    "torchvision==0.23.0",
    "lightning==2.5.1",
    "torchmetrics==1.7.2",
    "hydra-core==1.3.2",
    "hydra-colorlog==1.2.0",
    "hydra-optuna-sweeper==1.2.0",
    "avalanche-lib==0.6.0",
    "librosa==0.9.2",
    "numba==0.56.4",
    "numpy==1.23.5",
    "scipy==1.13.1",
    "rootutils==1.0.7",
    "pre-commit==4.2.0",
    "rich==14.0.0",
    "pytest==8.4.0",
    "pandas==2.2.3",
    "matplotlib==3.9.4",
    "tensorboard==2.19.0",
    "wandb==0.19.10",
    "peft==0.15.1",
    "transformers==4.51.3",
    "soundfile==0.13.1",
    "audiomentations==0.40.0",
    "pydub==0.25.1",
    "praat-parselmouth==0.4.5",
    "psola==0.0.1",
    "python-stretch==0.3.1",
    "numpy-minmax==0.3.1",
    "numpy-rms==0.4.2",
    "einops==0.8.1",
    "speechbrain==1.0.1",
    "accelerate==1.7.0",
    "huggingface-hub==0.32.4",
    "hyperpyyaml==1.2.2",
    "lightning-utilities==0.14.3",
    "omegaconf==2.3.0",
    "optuna==2.10.1",
    "pydantic==2.11.5",
    "pydantic-core==2.33.2",
    "safetensors==0.5.3",
    "scikit-learn==1.6.1",
    "sentencepiece==0.2.0",
    "tokenizers==0.21.1",
    "triton==3.4.0",
    "espnet>=202506",
    "boto3==1.35.36",
    "dvc[s3]==3.55.2",
    "gitpython==3.1.43",
    "cloudml-hypertune>=0.1.0.dev6",
]
